+++ test/analytics.test.js
@@ -111,6 +111,36 @@ describe('Analytics', function() {
       group.load.restore();
     });
 
+    it('should gracefully handle integrations that fail to initialize', function() {
+      Test.prototype.initialize = function() { throw new Error('Uh oh!'); };
+      analytics.add(Test);
+      analytics.initialize();
+      assert(analytics.initialized);
+    });
+
+    it('should store the names of integrations that did not initialize', function() {
+      Test.prototype.initialize = function() { throw new Error('Uh oh!'); };
+      analytics.add(Test);
+      analytics.initialize();
+      assert(analytics.failedInitializations.length === 1);
+      assert(analytics.failedInitializations[0] === Test.prototype.name);
+    });
+
+    it('should handle cases where integration.initialize failes and integration.prototype is undefined', function() {
+      Test.initialize = function() { throw new Error('Uh oh!'); };
+      Test.prototype = undefined;
+      analytics.add(Test);
+      analytics.initialize();
+      assert(analytics.failedInitializations[0] === 'Unknown');
+    });
+
+    it('should handle cases where initialization fails and integration.prototype.name is undefined', function() {
+      Test.prototype.name = undefined;
+      analytics.add(Test);
+      analytics.initialize();
+      assert(analytics.failedInitializations[0] === 'Unknown');
+    });
+
     it('should not error without settings', function() {
       analytics.initialize();
     });