+++ HISTORY.md
@@ -1,3 +1,10 @@
+3.1.1 / 2017-10-31
+==================
+
+  * Wrap initialize functions of integrations in try/catch statement.
+  * Add logging of failed initializations.
+  * Add a `failedInitializations` array to prototype to capture names of any failed integrations.
+
 3.1.0 / 2017-06-29
 ==================